Moved WhoAmI to core-plugins and added service information
