+++ muikku-rest/src/main/java/fi/otavanopisto/muikku/rest/user/UserRESTService.java
@@ -1,6 +1,5 @@
 package fi.otavanopisto.muikku.rest.user;
 
-import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -10,7 +9,6 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
@@ -45,24 +43,18 @@ import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.time.DateUtils;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import fi.otavanopisto.muikku.controller.PermissionController;
 import fi.otavanopisto.muikku.controller.SystemSettingsController;
 import fi.otavanopisto.muikku.dao.base.SchoolDataSourceDAO;
 import fi.otavanopisto.muikku.dao.users.UserPendingPasswordChangeDAO;
 import fi.otavanopisto.muikku.i18n.LocaleController;
 import fi.otavanopisto.muikku.mail.Mailer;
 import fi.otavanopisto.muikku.model.base.SchoolDataSource;
-import fi.otavanopisto.muikku.model.security.Permission;
 import fi.otavanopisto.muikku.model.users.EnvironmentRoleArchetype;
 import fi.otavanopisto.muikku.model.users.EnvironmentRoleEntity;
 import fi.otavanopisto.muikku.model.users.Flag;
 import fi.otavanopisto.muikku.model.users.FlagShare;
 import fi.otavanopisto.muikku.model.users.FlagStudent;
 import fi.otavanopisto.muikku.model.users.OrganizationEntity;
-import fi.otavanopisto.muikku.model.users.UserEmailEntity;
 import fi.otavanopisto.muikku.model.users.UserEntity;
 import fi.otavanopisto.muikku.model.users.UserEntityProperty;
 import fi.otavanopisto.muikku.model.users.UserGroupEntity;
@@ -80,16 +72,13 @@ import fi.otavanopisto.muikku.rest.model.Student;
 import fi.otavanopisto.muikku.rest.model.StudentAddress;
 import fi.otavanopisto.muikku.rest.model.StudentEmail;
 import fi.otavanopisto.muikku.rest.model.StudentPhoneNumber;
-import fi.otavanopisto.muikku.rest.model.UserWhoAmIInfo;
 import fi.otavanopisto.muikku.schooldata.BridgeResponse;
-import fi.otavanopisto.muikku.schooldata.CourseMetaController;
 import fi.otavanopisto.muikku.schooldata.GradingController;
 import fi.otavanopisto.muikku.schooldata.RestCatchSchoolDataExceptions;
 import fi.otavanopisto.muikku.schooldata.SchoolDataBridgeSessionController;
 import fi.otavanopisto.muikku.schooldata.SchoolDataIdentifier;
 import fi.otavanopisto.muikku.schooldata.UserSchoolDataController;
 import fi.otavanopisto.muikku.schooldata.WorkspaceEntityController;
-import fi.otavanopisto.muikku.schooldata.entity.Curriculum;
 import fi.otavanopisto.muikku.schooldata.entity.GradingScale;
 import fi.otavanopisto.muikku.schooldata.entity.GradingScaleItem;
 import fi.otavanopisto.muikku.schooldata.entity.StudyProgramme;
@@ -106,7 +95,6 @@ import fi.otavanopisto.muikku.search.SearchResults;
 import fi.otavanopisto.muikku.security.MuikkuPermissions;
 import fi.otavanopisto.muikku.security.RoleFeatures;
 import fi.otavanopisto.muikku.servlet.BaseUrl;
-import fi.otavanopisto.muikku.session.CurrentUserSession;
 import fi.otavanopisto.muikku.session.SessionController;
 import fi.otavanopisto.muikku.users.CreatedUserEntityFinder;
 import fi.otavanopisto.muikku.users.FlagController;
@@ -189,9 +177,6 @@ public class UserRESTService extends AbstractRESTService {
   private OrganizationEntityController organizationEntityController;
   
   @Inject
-  private PermissionController permissionController;
-
-  @Inject
   private UserPendingPasswordChangeDAO userPendingPasswordChangeDAO;
   
   @Inject
@@ -201,12 +186,6 @@ public class UserRESTService extends AbstractRESTService {
   @Any
   private Instance<SearchProvider> searchProviders;
 
-  @Inject
-  private CurrentUserSession currentUserSession;
-  
-  @Inject 
-  private CourseMetaController courseMetaController;
-  
   @Inject
   private UserEmailEntityController userEmailEntityController;
 
@@ -1779,137 +1758,6 @@ public class UserRESTService extends AbstractRESTService {
         .cacheControl(cacheControl)
         .tag(tag)
         .build();
-  }
-
-  @GET
-  @Path("/whoami")
-  @RESTPermit(handling = Handling.INLINE)
-  public Response findWhoAmI(@Context Request request) {
-    UserEntity userEntity = sessionController.getLoggedUserEntity();
-    SchoolDataIdentifier userIdentifier = userEntity == null ? null : sessionController.getLoggedUser();
-
-    // User role
-    
-    EnvironmentRoleArchetype role = null;
-    if (userIdentifier != null) {
-      UserSchoolDataIdentifier userSchoolDataIdentifier = userSchoolDataIdentifierController.
-          findUserSchoolDataIdentifierBySchoolDataIdentifier(userIdentifier);
-      role = userSchoolDataIdentifier.getRole().getArchetype();
-    }
-    
-    // Environment level permissions
-    
-    Set<String> permissionSet = new HashSet<String>();
-    List<Permission> permissions = permissionController.listPermissionsByScope("ENVIRONMENT");
-    for (Permission permission : permissions) {
-      if (sessionController.hasEnvironmentPermission(permission.getName())) {
-        permissionSet.add(permission.getName());
-      }
-    }
-    
-    // Response
-    
-    User user = userIdentifier == null ? null : userController.findUserByIdentifier(userIdentifier);
-    
-    String organizationIdentifier = user == null ? null : user.getOrganizationIdentifier().toId();
-    String defaultOrganizationIdentifier = systemSettingsController.getSetting("defaultOrganization");
-    boolean isDefaultOrganization = user == null ? false : StringUtils.equals(organizationIdentifier,  defaultOrganizationIdentifier);
-    boolean hasImage = userEntity == null ? false : userEntityFileController.hasProfilePicture(userEntity);
-    
-    // Study dates
-
-    OffsetDateTime studyStartDate = user == null ? null : user.getStudyStartDate();
-    OffsetDateTime studyEndDate = user == null ? null : user.getStudyEndDate();
-    OffsetDateTime studyTimeEnd = user == null ? null : user.getStudyTimeEnd();
-    String studyTimeLeftStr = userEntityController.getStudyTimeEndAsString(studyTimeEnd);
-    
-    // Curriculum 
-    String curriculumName = null;
-    
-    if (user != null) {
-      if (user.getCurriculumIdentifier() != null) {
-        SchoolDataIdentifier curriculumId = user.getCurriculumIdentifier();
-      
-        Curriculum curriculum = courseMetaController.findCurriculum(curriculumId.getDataSource(), curriculumId.getIdentifier());
-        curriculumName = curriculum == null ? null : curriculum.getName();
-      }
-    }
-    // Damn emails, addresses, and phoneNumbers as json
-    
-    String emails = null;
-    if (userIdentifier != null) {
-      List<String> foundEmails = userEmailEntityController.getUserEmailAddresses(userIdentifier).stream().map(UserEmailEntity::getAddress).collect(Collectors.toList());
-      try {
-        emails = new ObjectMapper().writeValueAsString(foundEmails);
-      }
-      catch (JsonProcessingException e) {
-        emails = null;
-      }
-    }
-    
-    String addresses = null;
-    if (userIdentifier != null) {
-      List<UserAddress> userAddresses = userController.listUserAddresses(user);
-      ArrayList<String> foundAddresses = new ArrayList<>();
-      for (UserAddress userAddress : userAddresses) {
-        foundAddresses.add(String.format("%s %s %s %s", userAddress.getStreet(), userAddress.getPostalCode(),
-            userAddress.getCity(), userAddress.getCountry()));
-      }
-      try {
-        addresses = new ObjectMapper().writeValueAsString(foundAddresses);
-      } 
-      catch (JsonProcessingException e) {
-        addresses = null;
-      }
-    }
-    
-    String phoneNumbers = null;
-    if (userIdentifier != null) {
-      List<UserPhoneNumber> userPhoneNumbers = userIdentifier == null ? null :  userController.listUserPhoneNumbers(user);
-      ArrayList<String> foundPhoneNumbers = new ArrayList<>();
-      for (UserPhoneNumber userPhoneNumber : userPhoneNumbers) {
-        foundPhoneNumbers.add(userPhoneNumber.getNumber());
-      }
-      try {
-        phoneNumbers = new ObjectMapper().writeValueAsString(foundPhoneNumbers);
-      }
-      catch (JsonProcessingException e) {
-        phoneNumbers = null;
-      }
-    }
-    
-    Locale localeObj = sessionController.getLocale();
-    String locale = (localeObj == null || localeObj.getLanguage() == null) ? "fi" : localeObj.getLanguage().toLowerCase();
-
-    // Result object
-    UserWhoAmIInfo whoamiInfo = new UserWhoAmIInfo(
-        userEntity == null ? null : userEntity.getId(),
-        userEntity == null ? null : userEntity.defaultSchoolDataIdentifier().toId(),
-        user == null ? null : user.getFirstName(),
-        user == null ? null : user.getLastName(),
-        user == null ? null : user.getNickName(),
-        user == null ? null : user.getStudyProgrammeName(),
-        user == null || user.getStudyProgrammeIdentifier() == null ? null : user.getStudyProgrammeIdentifier().toId(),
-        hasImage,
-        user == null ? false : user.getHasEvaluationFees(),
-        user == null || user.getCurriculumIdentifier() == null ? null : user.getCurriculumIdentifier().toId(),
-        curriculumName,
-        organizationIdentifier,
-        isDefaultOrganization,
-        currentUserSession.isActive(),
-        permissionSet,
-        role,
-        locale,
-        user == null ? null : user.getDisplayName(),
-        emails,
-        addresses,
-        phoneNumbers,
-        studyTimeLeftStr,
-        studyStartDate,
-        studyEndDate,
-        studyTimeEnd); 
-
-    return Response.ok(whoamiInfo).build();
   }
 
   @GET
